adicionando Dockerfile ao projeto

+++ src/database/migrations/00000000000003_order_product.ts
@@ -7,7 +7,7 @@ export async function up(knex: Knex): Promise<void> {
 
         table.index(['orderId', 'productId']);
 
-        table.decimal('value').notNullable()
+        table.double('value').notNullable()
     })
 }
